+++ docs/changes.rst
@@ -14,9 +14,10 @@ Version 0.5.6
 
 Unreleased.
 
- - Added :meth:`Image.percent_escape() <wand.image.BaseImage.percent_escape>` helper method. [:issue:`421`]
  - Added :meth:`Image.annotate() <wand.image.BaseImage.annotate>` method. [:issue:`#418`]
-
+ - Added :meth:`Image.percent_escape() <wand.image.BaseImage.percent_escape>` helper method. [:issue:`421`]
+ - Added :meth:`Image.ping() <wand.image.Image.ping>` class method. [:issue:`#425`]
+ 
 
 .. _changelog-0.5.5: