+++ docs/changes.rst
@@ -16,6 +16,7 @@ Unreleased.
 
  - Fixed invalid escape sequence warnings [:issue:`#428`]
  - Fixed error on Drawing exception handling. [:issue:`#427`]
+ - Fixed undefined behaviour when working with image frames in ImageMagick-7. [:issue:`#431`]
  - Added :meth:`Image.annotate() <wand.image.BaseImage.annotate>` method. [:issue:`#418`]
  - Added :meth:`Image.level_colors() <wand.image.BaseImage.level_colors>` method.
  - Added :meth:`Image.levelize_colors() <wand.image.BaseImage.levelize_colors>` method.